Update deploy scripts.

+++ start_node_genesis_gcp.sh
@@ -1,40 +1,70 @@
 #!/bin/bash
 
-if [[ "$#" -lt 3 ]]; then
-    echo "Usage: bash start_node_genesis_gcp.sh [dev|staging|spring|summer] <Shard Index> <Node Index> [--keystore]"
+if [[ "$#" -lt 3 ]] || [[ "$#" -gt 5 ]]; then
+    echo "Usage: bash start_node_genesis_gcp.sh [dev|staging|spring|summer] <Shard Index> <Node Index> [--keystore] [--keep-code]"
     echo "Example: bash start_node_genesis_gcp.sh spring 0 0 --keystore"
     exit
 fi
-OPTIONS=$4
-echo "OPTIONS=$OPTIONS"
 
+KEEP_CODE=""
+if [[ "$#" -gt 3 ]]; then
+    if [[ "$4" = '--keystore' ]]; then
+        OPTIONS=$4
+    elif [[ "$4" = '--keep-code' ]]; then
+        KEEP_CODE=true
+    else
+        printf "Invalid option: $4\n"
+        exit
+    fi
+fi
+if [[ "$#" = 5 ]]; then
+    if [[ "$5" = '--keystore' ]]; then
+        OPTIONS=$5
+    elif [[ "$5" = '--keep-code' ]]; then
+        KEEP_CODE=true
+    else
+        printf "Invalid option: $5\n"
+        exit
+    fi
+fi
+echo "OPTIONS=$OPTIONS"
+echo "KEEP_CODE=$KEEP_CODE"
 
 echo 'Killing old jobs..'
 sudo killall node
 
+if [[ "$KEEP_CODE" = "" ]]; then
+    echo 'Setting up working directory..'
+    cd
+    sudo rm -rf /home/ain_blockchain_data
+    sudo mkdir /home/ain_blockchain_data
+    sudo chmod -R 777 /home/ain_blockchain_data
+    sudo rm -rf ../ain-blockchain*
+    sudo mkdir ../ain-blockchain
+    sudo chmod -R 777 ../ain-blockchain
+    mv * ../ain-blockchain
+    cd ../ain-blockchain
 
-echo 'Setting up working directory..'
-cd
-sudo rm -rf /home/ain_blockchain_data
-sudo mkdir /home/ain_blockchain_data
-sudo chmod 777 /home/ain_blockchain_data
-sudo rm -rf ../ain-blockchain*
-sudo mkdir ../ain-blockchain
-sudo chmod 777 ../ain-blockchain
-mv * ../ain-blockchain
-cd ../ain-blockchain
+    echo 'Installing node modules..'
+    npm install
+else
+    echo 'Using old directory..'
+    OLD_DIR_PATH=$(find ../ain-blockchain* -maxdepth 0 -type d)
+    printf "OLD_DIR_PATH=$OLD_DIR_PATH\n"
+    sudo chmod -R 777 $OLD_DIR_PATH
+    sudo chmod -R 777 /home/ain_blockchain_data
+    cd $OLD_DIR_PATH
+fi
 
 
-echo 'Installing node modules..'
-npm install
 export GENESIS_CONFIGS_DIR=genesis-configs/testnet
 KEYSTORE_DIR=testnet_dev_staging_keys
 if [[ "$1" = 'spring' ]]; then
     export TRACKER_WS_ADDR=ws://35.221.137.80:5000
-    KEYSTORE_DIR=testnet_spring_summer_keys
+    KEYSTORE_DIR=testnet_prod_keys
 elif [[ "$1" = 'summer' ]]; then
     export TRACKER_WS_ADDR=ws://35.194.172.106:5000
-    KEYSTORE_DIR=testnet_spring_summer_keys
+    KEYSTORE_DIR=testnet_prod_keys
 elif [[ "$1" = 'staging' ]]; then
     export TRACKER_WS_ADDR=ws://35.221.150.73:5000
 elif [[ "$1" = 'dev' ]]; then
@@ -106,32 +136,25 @@ fi
 if [[ "$OPTIONS" != '--keystore' ]] || [[ "$2" -gt 0 ]]; then
     export ACCOUNT_INDEX="$3"
     echo "ACCOUNT_INDEX=$ACCOUNT_INDEX"
-    COMMAND_PREFIX=""
 else
     if [[ "$3" = 0 ]]; then
-        KEYSTORE_FILENAME="2021-09-24T00-00-00"
+        KEYSTORE_FILENAME="keystore_node_0.json"
     elif [[ "$3" = 1 ]]; then
-        KEYSTORE_FILENAME="2021-09-24T00-00-01"
+        KEYSTORE_FILENAME="keystore_node_1.json"
     elif [[ "$3" = 2 ]]; then
-        KEYSTORE_FILENAME="2021-09-24T00-00-02"
+        KEYSTORE_FILENAME="keystore_node_2.json"
     elif [[ "$3" = 3 ]]; then
-        KEYSTORE_FILENAME="2021-09-24T00-00-03"
+        KEYSTORE_FILENAME="keystore_node_3.json"
     else
-        KEYSTORE_FILENAME="2021-09-24T00-00-04"
+        KEYSTORE_FILENAME="keystore_node_4.json"
     fi
     echo "KEYSTORE_FILENAME=$KEYSTORE_FILENAME"
-    sudo mkdir -p ../ain_blockchain_data/keys/8080
-    sudo mv ./$KEYSTORE_DIR/$KEYSTORE_FILENAME ../ain_blockchain_data/keys/8080/
-
-    # Set up a pipe named `/tmp/blockchain_node_fifo`
-    mkfifo /tmp/blockchain_node_fifo
-
-    # To avoid your server receiving an EOF. At least one process must have
-    # the fifo opened in writing so your server does not receive a EOF.
-    cat > /tmp/blockchain_node_fifo &
-
-    # Starts the server reading from the pipe named `/tmp/blockchain_node_fifo`
-    COMMAND_PREFIX="cat /tmp/blockchain_node_fifo |"
+    if [[ "$KEEP_CODE" = "" ]]; then
+        sudo mkdir -p ../ain_blockchain_data/keys/8080
+        sudo mv ./$KEYSTORE_DIR/$KEYSTORE_FILENAME ../ain_blockchain_data/keys/8080/
+    fi
+    export KEYSTORE_FILE_PATH=/home/ain_blockchain_data/keys/8080/$KEYSTORE_FILENAME
+    echo "KEYSTORE_FILE_PATH=$KEYSTORE_FILE_PATH"
 fi
 
 export DEBUG=false
@@ -145,7 +168,7 @@ export BLOCKCHAIN_DATA_DIR="/home/ain_blockchain_data"
 MAX_OLD_SPACE_SIZE_MB=6000
 
 printf "\nStarting up Blockchain Node server..\n\n"
-START_CMD="$COMMAND_PREFIX nohup node --async-stack-traces --max-old-space-size=$MAX_OLD_SPACE_SIZE_MB client/index.js >/dev/null 2>error_logs.txt &"
+START_CMD="nohup node --async-stack-traces --max-old-space-size=$MAX_OLD_SPACE_SIZE_MB client/index.js >/dev/null 2>error_logs.txt &"
 printf "START_CMD='$START_CMD'\n"
 eval $START_CMD
